refactor zoom components remove char card
