+++ src/components/organizational-chart/OrganizationalChart.jsx
@@ -1,7 +1,10 @@
-import {useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { GetTiersList } from "../../common/functions/org-chart-functions";
 import D3TreeOrgChart from "./d3-tree-org-chart/D3TreeOrgChart";
 import { OrganizationalData } from "../../common/mock/organizational-data";
+import { Box } from "@mui/material";
+import TierList from "./TierList";
+import ZoomControls from "../../common/components/org-chart/ZoomControls";
 
 const OrganizationalChart = () => {
   const chartData = {
@@ -16,7 +19,7 @@ const OrganizationalChart = () => {
   const [treeData, setTreeData] = useState(chartData);
   const [nodeTiers, setNodeTiers] = useState(0);
   const [isLoading, setIsLoading] = useState(false);
-
+  const [zoom, setZoom] = useState(1);
 
   useEffect(() => {
     const loadChartData = async () => {
@@ -38,7 +41,18 @@ const OrganizationalChart = () => {
   }, []);
 
   return (
-    <D3TreeOrgChart />
+    <Box
+      sx={{
+        width: "100%",
+        height: "100vh",
+        backgroundColor: "#f5f5f5",
+        position: "relative",
+      }}
+    >
+      <TierList tiers={nodeTiers} />
+      <ZoomControls  onChangeZoom={(value) => setZoom(value)} />
+      {!isLoading && <D3TreeOrgChart treeData={treeData}  zoom={zoom}/>}
+    </Box>
   );
 };
